requirements.txt
Flask>=2.3.0
openai-whisper>=20231117
pydub>=0.25.1
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg
